backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pymongo>=4.9,<4.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.18
pydantic[email]==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
email-validator